Properly throw error in processPixelResults
If the result row for pixel query fails and an error is returned, the results object will be null or undefined. However, processPixelResults attempts to read properties from the results object before the error can be propagated to the calling function. This results in an "attempt to access property of undefined" error, missing the original error returned from the server.  This change provides an early out, throwing or returning the error immediately so the caller can understand what went wrong.