Don't strip precision from row descriptors
In the table create service, the helper function `mutateThriftRowDesc()` was stripping out the `precision` property of a row descriptor. This was causing problems with the PR mapd/mapd-immerse#4920, which adds support for setting higher-precision timestamps (ms, us, ns) in the table importer. The bug in mapd-connector meant that, even though the precision of the column was being set to a non-zero value, that precision wasn't getting passed to the backend.

This commit fixes the bug by ensuring the precision property of a row descriptor is retained and passed to the backend intact. This allows us to successfully set the precision in Immerse, and have the backend respect that precision.