Fix export/import error with `timestampToMs()`
When I added the `timestampToMs()` helper function to `utils.js`, I also switched the export style from `module.exports` to `export const ...`, because I am a hip and modern guy and I am totally down with the kids and their es6. But it turns out I am just an es5 guy living in an es6 world, and I messed this up.

I think I've fixed the bug in this commit by destructuring the import when we need it.

I also discovered in my debugging that the only other function in `utils.js`, `bindArgsFromN()`, isn't ever used anywhere and is dead code. I went ahead and removed it. And having removed it, my `timestampToMs()` was the only thing left in that file, so I decided to move it to `helpers.js`. Going by name alone, these two files seem to have the same exact purpose, and I imagine the only reason we have two of them is because of cruft. But now I moved my code to `helpers` and deleted `utils`. Yay for less cruft.