Properly handle timestamps with precision >0 [FE-5229] (#105)
The DB stores timestamps as integer values. Depending on the precision of the column, this int value may represent seconds (precision=0), milliseconds (p=3), microseconds (p=6) or nanoseconds (p=9) since the epoch. When a timestamp value is retrieved from the DB, we convert it to a JS Date object. The constructor for Date takes the time in ms since the epoch.

Previously, we assumed that all timestamp values in the DB were precision 0 (seconds), and so simply multiplied that by `1000` to get ms, and passed that to `new Date()`. This commit modifies that behavior to properly handle timestamps of differing precisions.

To accomplish this, we check the precision of the column of the value. We use this to calculate a divisor for the integer value that will convert it to ms. For example, if the precision is 0, we get `10^(0 - 3)=10^(-3)=0.001`. Dividing 2 seconds by 0.001 gives us 2000, the correct value in ms.

N.B.: the JS Date object will only have ms-level precision, regardless of the precision of column in the DB. This is a JS limitation. Any precision beyond ms will be discarded by the connector.