Split off client init and client connect into separate functions
This split will allow us to implement SAML support in Immerse. In that case, the web server is the one that makes the `connect` call to the DB, using the SAML credentials it receives, and returns a valid DB session ID to Immerse. This means that we shouldn't (and, indeed, can not) call `connect` from connector.

By splitting all the non `connect` logic into the `init` function, we can still create and initialize our connector clients, but stop short of actually calling `connect`. The original `connect` call should call `init` *and* then call `connect`, so its behavior should remain unchanged.