Don't require a db name to call `initClients()` + have it return `this`
This is in support of our SAML flow, which doesn't know the username, password, or database when connecting to the database (that is handled in our server-side SAML authentication mechanism). Previously, `initClients()` would check for the presence of a database name and, if missing, throw an error. Now it no longer checks this property, and is perfectly happy to run without a database name being set. (This error checking logic has been added to `connect()`, which needs all this info during its normal 'ol authentication method).

Additionally, `initClients()` now returns `this`, so that additional connector method calls can be chained off of it. `connect()`, which calls `initClients()`, has been updated to account for this.