[FE-7671] Split client init and connect into separate functions, expose get_session_info for SAML support (#131)
* Split off client init and client connect into separate functions

This split will allow us to implement SAML support in Immerse. In that case, the web server is the one that makes the `connect` call to the DB, using the SAML credentials it receives, and returns a valid DB session ID to Immerse. This means that we shouldn't (and, indeed, can not) call `connect` from connector.

By splitting all the non `connect` logic into the `init` function, we can still create and initialize our connector clients, but stop short of actually calling `connect`. The original `connect` call should call `init` *and* then call `connect`, so its behavior should remain unchanged.

* Fix handling of client init in Node.js mode (fixes failing unit tests)

* Expose the `get_session_info` Thrift service as `getSessionInfoAsync()`

* Update Thrift

* Fix definition of getSessionInfoAsync()

* Wrap getSessionInfoAsync in handleErrors

* Compile for newest changes

Who's smart idea was it to compile javascript in the first place?

* Compile with latest Thrift bindings

* arrayify sessionId

* remove handleErrors (possibly revert?)

* set _this.client

* compile last three commits

* Don't require a db name to call `initClients()` + have it return `this`

This is in support of our SAML flow, which doesn't know the username, password, or database when connecting to the database (that is handled in our server-side SAML authentication mechanism). Previously, `initClients()` would check for the presence of a database name and, if missing, throw an error. Now it no longer checks this property, and is perfectly happy to run without a database name being set. (This error checking logic has been added to `connect()`, which needs all this info during its normal 'ol authentication method).

Additionally, `initClients()` now returns `this`, so that additional connector method calls can be chained off of it. `connect()`, which calls `initClients()`, has been updated to account for this.

* make isTimeoutError a method so we can use it outside of connector

* Add usingSAML flag

* re-implement handleErrors on getSessionInfo

* compile

* Rename `usingSaml()` to `disableReconnect()`

There wasn't anything SAML-specific in this flag, it just disabled the auto-reconnect logic when encountering a connection error (like a timeout). This flag just happens to be necessary for SAML, but not specific to it. This rename should make that more clear.

* Fix typo: `disableReconnect` -> `disableAutoReconnect`

* Add/fix jsdoc comments for new functions

* Run `eslint --fix ./src/` to auto-format source

* Fix a bug that caused duplicate clients

* fix num connections for init vs connect (#137)

* testing moving _numConnections to init not connect

* adding dist