Fix a bug that prevented connector from working properly under Node.js
The Node.js Thrift bindings use a different callback format than the browser-flavored JS version: `fn(err, results) {...}` instead of `fn(results)`. Connector wasn't aware of this, and so would only look at the first argument to the callback when passing back function call results to the caller. When there was no error, this would simply be `null`, leading to a bug where all functions seemingly refused to ever return data.

This commit specializes the `promisifyThriftMethod` wrapper function for both the browser and node.js. The browser version is the same as the existing function. The node.js version changes the callback arg format to what's described above. The runtime then dynamically chooses one specialization or the other based on the `isNodeRuntime()` function.