[FE-8720] Fix connector for Node.js & bump version to 4.7.2 (#157)
* Fix a bug that prevented connector from working properly under Node.js

The Node.js Thrift bindings use a different callback format than the browser-flavored JS version: `fn(err, results) {...}` instead of `fn(results)`. Connector wasn't aware of this, and so would only look at the first argument to the callback when passing back function call results to the caller. When there was no error, this would simply be `null`, leading to a bug where all functions seemingly refused to ever return data.

This commit specializes the `promisifyThriftMethod` wrapper function for both the browser and node.js. The browser version is the same as the existing function. The node.js version changes the callback arg format to what's described above. The runtime then dynamically chooses one specialization or the other based on the `isNodeRuntime()` function.

* Add tests for `getDashboardsAsync()` under node.js

This specifically tests that the bug fix for FE-8720 is working as intended. Note that the new test relies on the list of dashboards on Metis (the server) not changing. If they do change, this test will suddenly start breaking.

* Tell eslint to ignore the thrift, build, and dist directories

* Update Node.js example to include example of `getDashboardsAsync()`

* Bump version: 4.7.0 -> 4.7.2