Add arrow wire support, Use nodejs thrift bindings
This adds support for the arrow over the wire transport
method added in OmniSci 5.4.

Additionally, it switches to using the nodejs generated thrift
bindings in the browser rather than the pure JS browser generated ones.

The nodejs bindings have the benefit of generating error-handling code
that we were doing ourselves, and brings support for promise OR async
support with callbacks OR Promises. The latter of which we have manually
wrapped many endpoints to support.