feat: Fix conversion of buffer to Boolean
Calling the Boolean constructor on a Buffer object always returns true regardless if the value is true or not.  This is noticeable when
Creating a query using the ST_CONTAINS function as a column in a select statement.
For example,

SELECT id, name ST_CONTAINS(omnisci_geo, ST_GeomFromText(‘POINT(-87.3592959471789 35.0011799880309)’, 4326) from omnisci_states

Should return true for only Tennessee however, it returns true for each state